refactor: react-router의 라우팅 설정을 별도 파일로 분리
