--- src/constants/endpoint.ts
@@ -1,4 +1,0 @@
-export const ENDPOINT = {
-  shoppingCart: '/react-shopping-cart',
-  confirmOrder: '/react-shopping-cart/confirm-order',
-} as const;
