+++ src/routes/router.constants.ts
@@ -1,0 +1,5 @@
+export const ENDPOINT = {
+  baseUrl: '/react-shopping-cart',
+  shoppingCart: '/',
+  confirmOrder: '/confirm-order',
+} as const;
